requirements.txt
streamlit==1.28.1
boto3==1.34.34
langchain-aws==0.1.0
langchain-community==0.0.13
langchain-postgres==0.0.3
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pypdf==4.0.1
feedparser==6.0.10
pandas==2.1.4
python-dotenv==1.0.0
tiktoken==0.5.2